# backend/rfc-service/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  description: API для системы управления запросами на изменение (RFC)
  title: RFC Management System API
  version: 1.0.0
servers:
- url: /
security:
- bearerAuth: []
paths:
  /user:
    get:
      description: |
        Возвращает список пользователей с поддержкой пагинации и поиска.
        Поиск осуществляется по ID, username, firstName и lastName (case-insensitive).
      operationId: getUsers
      parameters:
      - description: Номер страницы (начиная с 0)
        in: query
        name: page
        required: false
        schema:
          default: 0
          minimum: 0
          type: integer
      - description: Размер страницы
        in: query
        name: size
        required: false
        schema:
          default: 20
          maximum: 100
          minimum: 1
          type: integer
      - description: "Строка поиска (по ID, username, firstName, lastName)"
        in: query
        name: searchString
        required: false
        schema:
          minLength: 1
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPageResponse"
          description: Успешный возврат списка пользователей
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal Server Error
      summary: Получить пагинированный список пользователей
      tags:
      - Users
    post:
      description: Создание нового пользователя в системе
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
          description: Пользователь успешно создан
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Unauthorized
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal Server Error
      summary: Создать нового пользователя
      tags:
      - Users
  /user/{id}:
    delete:
      description: Удаление пользователя из системы по идентификатору
      operationId: deleteUser
      parameters:
      - description: ID элемента
        in: path
        name: id
        required: true
        schema:
          format: int64
          nullable: false
          type: integer
      responses:
        "204":
          description: Пользователь успешно удален
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal Server Error
      summary: Удалить пользователя
      tags:
      - Users
    get:
      description: Возвращает детальную информацию о пользователе по его идентификатору
      operationId: getUserById
      parameters:
      - description: ID элемента
        in: path
        name: id
        required: true
        schema:
          format: int64
          nullable: false
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
          description: Успешный возврат пользователя
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal Server Error
      summary: Получить пользователя по ID
      tags:
      - Users
    put:
      description: Полное обновление информации о пользователе
      operationId: updateUser
      parameters:
      - description: ID элемента
        in: path
        name: id
        required: true
        schema:
          format: int64
          nullable: false
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
          description: Пользователь успешно обновлен
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal Server Error
      summary: Обновить пользователя
      tags:
      - Users
  /user/me:
    get:
      description: Возвращает информацию о текущем авторизованном пользователе
      operationId: getCurrentUser
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
          description: Успешное получение информации о пользователе
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal Server Error
      summary: Получить информацию о текущем пользователе
      tags:
      - Users
  /user/login:
    post:
      description: "Аутентификация пользователя по имени и паролю, возвращает JWT\
        \ токены"
      operationId: loginUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
          description: Успешная аутентификация
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal Server Error
      summary: Аутентификация пользователя
      tags:
      - Users
  /team:
    get:
      description: Возвращает список команд с поддержкой пагинации и фильтрации по
        названию
      operationId: getTeams
      parameters:
      - description: Номер страницы (начиная с 0)
        in: query
        name: page
        required: false
        schema:
          default: 0
          minimum: 0
          type: integer
      - description: Размер страницы
        in: query
        name: size
        required: false
        schema:
          default: 20
          maximum: 100
          minimum: 1
          type: integer
      - description: Фильтр по названию команды (поиск по совпадению)
        in: query
        name: name
        required: false
        schema:
          minLength: 1
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamPageResponse"
          description: Успешный возврат списка команд
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal Server Error
      summary: Получить пагинированный список команд
      tags:
      - Teams
    post:
      description: Создание новой команды в системе
      operationId: createTeam
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamResponse"
          description: Команда успешно создана
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal Server Error
      summary: Создать новую команду
      tags:
      - Teams
  /team/{id}:
    delete:
      description: Удаление команды по идентификатору
      operationId: deleteTeam
      parameters:
      - description: ID элемента
        in: path
        name: id
        required: true
        schema:
          format: int64
          nullable: false
          type: integer
      responses:
        "204":
          description: Команда успешно удалена
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal Server Error
      summary: Удалить команду
      tags:
      - Teams
    get:
      description: Возвращает детальную информацию о команде по её идентификатору
      operationId: getTeamById
      parameters:
      - description: ID элемента
        in: path
        name: id
        required: true
        schema:
          format: int64
          nullable: false
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamResponse"
          description: Успешный возврат команды
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal Server Error
      summary: Получить команду по ID
      tags:
      - Teams
    put:
      description: Полное обновление информации о команде
      operationId: updateTeam
      parameters:
      - description: ID элемента
        in: path
        name: id
        required: true
        schema:
          format: int64
          nullable: false
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamResponse"
          description: Команда успешно обновлена
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal Server Error
      summary: Обновить команду
      tags:
      - Teams
...
components:
...
  schemas:
    UserPageResponse:
      allOf:
      - $ref: "#/components/schemas/PageResponse"
      - properties:
          content:
            items:
              $ref: "#/components/schemas/UserResponse"
            type: array
        required:
        - content
        type: object
      example:
        number: 5
        size: 1
        last: true
        totalPages: 6
        first: true
        content:
        - firstName: firstName
          lastName: lastName
          role: USER
          id: 5
          username: username
        - firstName: firstName
          lastName: lastName
          role: USER
          id: 5
          username: username
        totalElements: 0
    UserResponse:
      description: Данные пользователя
      example:
        firstName: firstName
        lastName: lastName
        role: USER
        id: 5
        username: username
      properties:
        id:
          description: Уникальный идентификатор пользователя
          format: int64
          type: integer
        username:
          description: Имя пользователя
          type: string
        firstName:
          description: Имя
          type: string
        lastName:
          description: Фамилия
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
      required:
      - firstName
      - id
      - lastName
      - role
      - username
      type: object
    ErrorResponse:
      example:
        errors:
        - code: code
          message: message
        - code: code
          message: message
      nullable: false
      properties:
        errors:
          items:
            $ref: "#/components/schemas/Error"
          type: array
      required:
      - errors
      type: object
    Error:
      example:
        code: code
        message: message
      properties:
        code:
          type: string
        message:
          type: string
      required:
      - code
      - message
      type: object
    UserRequest:
      description: Данные для создания пользователя
      example:
        firstName: firstName
        lastName: lastName
        password: password
        role: USER
        username: username
      properties:
        username:
          description: Имя пользователя (уникальное)
          maxLength: 50
          minLength: 3
          pattern: "^[a-zA-Z0-9_-]+$"
          type: string
        firstName:
          description: Имя
          maxLength: 100
          minLength: 1
          type: string
        lastName:
          description: Фамилия
          maxLength: 100
          minLength: 1
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        password:
          description: Пароль (обязательное поле при создании)
          format: password
          maxLength: 100
          minLength: 8
          type: string
      required:
      - firstName
      - lastName
      - password
      - role
      - username
      type: object
    UpdateUserRequest:
      description: Данные для обновления пользователя
      example:
        firstName: firstName
        lastName: lastName
        role: USER
        username: username
      properties:
        username:
          description: Имя пользователя (уникальное)
          maxLength: 50
          minLength: 3
          pattern: "^[a-zA-Z0-9_-]+$"
          type: string
        firstName:
          description: Имя
          maxLength: 100
          minLength: 1
          type: string
        lastName:
          description: Фамилия
          maxLength: 100
          minLength: 1
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
      required:
      - firstName
      - lastName
      - role
      - username
      type: object
    LoginRequest:
      description: Данные для входа в систему
      example:
        password: password
        username: username
      properties:
        username:
          description: Имя пользователя
          maxLength: 50
          minLength: 1
          type: string
        password:
          description: Пароль
          format: password
          maxLength: 100
          minLength: 1
          type: string
      required:
      - password
      - username
      type: object
    LoginResponse:
      description: Токены доступа после успешной аутентификации
      example:
        expiresIn: 0
        accessToken: accessToken
        tokenType: tokenType
        refreshToken: refreshToken
        refreshExpiresIn: 6
      properties:
        accessToken:
          description: JWT токен доступа
          type: string
        refreshToken:
          description: Refresh токен для обновления access токена
          type: string
        expiresIn:
          description: Время жизни токена в секундах
          type: integer
        refreshExpiresIn:
          description: Время жизни refresh токена в секундах
          type: integer
        tokenType:
          description: Тип токена (обычно "Bearer")
          type: string
      required:
      - accessToken
      - expiresIn
      - tokenType
      type: object
    TeamPageResponse:
      allOf:
      - $ref: "#/components/schemas/PageResponse"
      - properties:
          content:
            items:
              $ref: "#/components/schemas/TeamResponse"
            type: array
        required:
        - content
        type: object
      example:
        number: 5
        size: 1
        last: true
        totalPages: 6
        first: true
        content:
        - members:
          - firstName: firstName
            lastName: lastName
            role: USER
            id: 5
            username: username
          - firstName: firstName
            lastName: lastName
            role: USER
            id: 5
            username: username
          name: name
          description: description
          id: 5
        - members:
          - firstName: firstName
            lastName: lastName
            role: USER
            id: 5
            username: username
          - firstName: firstName
            lastName: lastName
            role: USER
            id: 5
            username: username
          name: name
          description: description
          id: 5
        totalElements: 0
    TeamResponse:
      description: Данные команды
      example:
        members:
        - firstName: firstName
          lastName: lastName
          role: USER
          id: 5
          username: username
        - firstName: firstName
          lastName: lastName
          role: USER
          id: 5
          username: username
        name: name
        description: description
        id: 5
      properties:
        id:
          description: Уникальный идентификатор команды
          format: int64
          type: integer
        name:
          description: Название команды
          type: string
        description:
          description: Описание команды
          nullable: true
          type: string
        members:
          description: Список членов команды
          items:
            $ref: "#/components/schemas/UserResponse"
          type: array
      required:
      - id
      - members
      - name
      type: object
    TeamRequest:
      description: Данные для создания/обновления команды
      example:
        name: name
        description: description
        memberIds:
        - 0
        - 0
      properties:
        name:
          description: Название команды
          maxLength: 255
          minLength: 1
          type: string
        description:
          description: Описание команды
          maxLength: 1000
          nullable: true
          type: string
        memberIds:
          description: Список ID членов команды
          items:
            format: int64
            type: integer
          minItems: 1
          type: array
      required:
      - memberIds
      - name
      type: object
...
    PageResponse:
      description: Общая схема для пагинированных ответов
      properties:
        totalElements:
          description: Общее количество элементов
          format: int64
          type: integer
        totalPages:
          description: Общее количество страниц
          type: integer
        size:
          description: Размер страницы
          type: integer
        number:
          description: Номер текущей страницы (начиная с 0)
          type: integer
        first:
          description: Является ли текущая страница первой
          type: boolean
        last:
          description: Является ли текущая страница последней
          type: boolean
      required:
      - first
      - last
      - number
      - size
      - totalElements
      - totalPages
      type: object
    UserRole:
      description: Роль пользователя в системе
      enum:
      - USER
      - RFC_APPROVER
      - CAB_MANAGER
      - ADMIN
      type: string
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      description: JWT token
      scheme: bearer
      type: http